Create notification model in util mudule
